tried to reduce computeMeanActivation complexity... half way done

+++ src/Util.cpp
@@ -98,7 +98,7 @@ void Util::computeImageClasses(vector<Image> &images, const CNNScheme &scheme, v
       imagesByClass.push_back(pair<string,vector<pair<string,Image *> > >(currentClass,newClass));
     }
   }
-  printf("Util::computeImageClasses::Found %u image classes\n",imagesByClass.size());
+  printf("Util::computeImageClasses::Found %u image classes\n",(unsigned int)imagesByClass.size());
   //Compute each image class and store it
   #pragma omp parallel for schedule(dynamic,1)
   for(vector<pair<string,vector<pair<string,Image *> > > >::iterator it = imagesByClass.begin(); it<imagesByClass.end(); it++){
@@ -111,6 +111,6 @@ void Util::computeImageClasses(vector<Image> &images, const CNNScheme &scheme, v
     #pragma omp critical (imageClasses)
     imageClasses.push_back(currentImageClass);
     #pragma omp critical (printf)
-    printf("Util::computeImageClasses::Done computing meanAct of image class %s based on %u images\n",it->first.c_str(),it->second.size());
+    printf("Util::computeImageClasses::Done computing meanAct of image class %s based on %u images\n",it->first.c_str(),(unsigned int)it->second.size());
   }
 }
